Impl ClassSerializer
