+++ gremlin-javascript/src/main/javascript/gremlin-javascript/test/unit/graphbinary/DateSerializer-test.js
@@ -22,4 +22,4 @@
  */
 'use strict';
 
-require('./DateSerializerTestSuite')({ ID: 0x04, name: 'Date' });
+require('./DateSerializerTestTemplate')({ ID: 0x04, name: 'Date' });
